refactor: 项目重构和代码清理
- 重构 AI 模块为独立子模块 (client, resolver, summarizer 等)
- 重构 parallel 模块 (branch, cow, graph, manager, merge)
- 新增 optimization 模块 (cache, compression, dedup, algorithms)
- 新增 query_optimizer 模块
- 新增 file_kv 子模块 (config, checkpoints, incremental, recovery)
- 添加 Prometheus metrics 支持
- 添加分布式协调 (etcd)
- 添加 WAL 和 PITR 恢复功能
- 新增示例和集成测试
- 清理文档，旧文档移至 archive
- 添加 .gitignore 忽略运行时生成的文件

+++ src/sparse_index.rs
@@ -526,7 +526,7 @@ mod tests {
             let index = manager.create_index(seg_id);
             for i in 0..150 {
                 let key = format!("key_{:04}", seg_id * 1000 + i);
-                index.maybe_add_index_point(&key, i as u64 * 100, i as u64);
+                index.maybe_add_index_point(&key, i * 100, i);
             }
             manager.save_index(seg_id).unwrap();
         }
